# docs/swagger.yaml
basePath: /service/api
definitions:
...
  dto.BaseResponse:
    properties:
      data: {}
      message_en:
        example: Success
        type: string
      message_th:
        example: สำเร็จ
        type: string
      status:
        example: success
        type: string
      status_code:
        example: 200
        type: integer
    type: object
...
  dto.CreatePayableDTO:
    properties:
      amount:
        description: จำนวนเงิน
        type: number
      balance:
        description: ยอดคงเหลือ
        type: number
      bank_id:
        description: รหัสบัญชีธนาคารที่เกี่ยวข้อง
        type: string
      due_date:
        description: วันที่ครบกำหนดชำระ
        type: string
      invoice_no:
        description: เลขที่ใบแจ้งหนี้
        type: string
      issue_date:
        description: วันที่ออกใบแจ้งหนี้
        type: string
      note:
        description: หมายเหตุ
        type: string
      payment_ref:
        description: เลขที่อ้างอิงการชำระเงิน
        type: string
      purchase_no:
        description: เลขที่ใบสั่งซื้อ
        type: string
      supplier:
        description: ชื่อผู้จำหน่าย
        type: string
    type: object
...
  dto.Pagination:
    properties:
      list:
        items: {}
        type: array
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      total_count:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  dto.PayableDTO:
    properties:
      amount:
        description: จำนวนเงิน
        type: number
      balance:
        description: ยอดคงเหลือ
        type: number
      bank_id:
        description: รหัสบัญชีธนาคารที่เกี่ยวข้อง
        type: string
      due_date:
        description: วันที่ครบกำหนดชำระ
        type: string
      id_payable:
        description: รหัสเจ้าหนี้
        type: string
      invoice_no:
        description: เลขที่ใบแจ้งหนี้
        type: string
      issue_date:
        description: วันที่ออกใบแจ้งหนี้
        type: string
      purchase_no:
        description: เลขที่ใบสั่งซื้อ
        type: string
      supplier:
        description: ชื่อผู้จำหน่าย
        type: string
    type: object
  dto.PayableSummaryDTO:
    properties:
      overdue_count:
        description: จำนวนรายการเกินกำหนด
        type: integer
      total_amount:
        description: ยอดรวมทั้งหมด
        type: number
      total_due:
        description: ยอดคงค้าง
        type: number
    type: object
...
  dto.RecordPaymentDTO:
    properties:
      amount:
        description: จำนวนเงินที่ชำระ
        type: number
      bank_id:
        description: รหัสบัญชีธนาคารที่รับชำระ
        type: string
      note:
        description: หมายเหตุเพิ่มเติม
        type: string
      payable_id:
        description: ไอดีของเจ้าหนี้/รายการที่ต้องชำระ
        type: string
      payment_date:
        description: วันที่ชำระเงิน (เช่น 2025-10-26)
        type: string
      payment_method:
        description: วิธีการชำระเงิน (เช่น โอนเงิน/เงินสด/เช็ค)
        type: string
      payment_ref:
        description: เลขอ้างอิงการชำระเงิน/หลักฐาน
        type: string
    type: object
...
  dto.UpdatePayableDTO:
    properties:
      amount:
        description: จำนวนเงิน
        type: number
      balance:
        description: ยอดคงเหลือ
        type: number
      bank_id:
        description: รหัสบัญชีธนาคารที่เกี่ยวข้อง
        type: string
      due_date:
        description: วันที่ครบกำหนดชำระ
        type: string
      invoice_no:
        description: เลขที่ใบแจ้งหนี้
        type: string
      issue_date:
        description: วันที่ออกใบแจ้งหนี้
        type: string
      note:
        description: หมายเหตุ
        type: string
      payment_ref:
        description: เลขที่อ้างอิงการชำระเงิน
        type: string
      purchase_no:
        description: เลขที่ใบสั่งซื้อ
        type: string
      status:
        description: สถานะ
        type: string
      supplier:
        description: ชื่อผู้จำหน่าย
        type: string
    type: object
...
host: api.dev.rkp-media.com
info:
  contact: {}
  description: This is an ERP API demo.
  title: ERP Demo API
  version: "1.0"
paths:
...
  /v1/payable/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a payable record by its ID
      parameters:
      - description: Payable ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Delete payable by ID
      tags:
      - Payables
    get:
      consumes:
      - application/json
      description: Retrieve a payable record by its ID
      parameters:
      - description: Payable ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PayableDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get payable by ID
      tags:
      - Payables
    put:
      consumes:
      - application/json
      description: Update a payable record by its ID
      parameters:
      - description: Payable ID
        in: path
        name: id
        required: true
        type: string
      - description: Payable data to update
        in: body
        name: payable
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePayableDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Update payable by ID
      tags:
      - Payables
  /v1/payable/create:
    post:
      consumes:
      - application/json
      description: Create a new payable record
      parameters:
      - description: Payable to create
        in: body
        name: payable
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePayableDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Create a new payable
      tags:
      - Payables
  /v1/payable/list:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of payables with optional search and
        filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - default: created_at
        description: Field to sort by
        enum:
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Pagination'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: List payables with pagination and filtering
      tags:
      - Payables
  /v1/payable/record-payment:
    post:
      consumes:
      - application/json
      description: Record a payment transaction and update the payable balance/status
        accordingly
      parameters:
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/dto.RecordPaymentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Record a payment for a payable
      tags:
      - Payables
  /v1/payable/summary:
    get:
      consumes:
      - application/json
      description: Retrieve a summary of payables based on filters
      parameters:
      - description: Bank ID
        in: query
        name: bank_id
        type: string
      - description: Report type
        enum:
        - day
        - month
        - all
        in: query
        name: report
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PayableSummaryDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseResponse'
      summary: Get payable summary by filter
      tags:
      - Payables
...
swagger: "2.0"